+++ packages/data/src/queries/generalSettings/useUpdateGeneralSettingsCache.ts
@@ -1,6 +1,6 @@
 import type { CacheUpdaterFn, WriteQueryOptions } from '../types';
 import useUpdateCache from '../useUpdateCache';
-import type { GeneralSettingsData } from '@eventespresso/services';
+import type { GeneralSettingsData } from '@eventespresso/data';
 
 const useUpdateGeneralSettingsCache = (
 	writeQueryOptions: WriteQueryOptions<GeneralSettingsData> = undefined
