+++ packages/edtr-services/src/types.ts
@@ -1,6 +1,6 @@
 import type { OptionsType } from '@eventespresso/adapters';
-import type { User } from '@eventespresso/data';
-import type { GeneralSettings, RelationalData } from '@eventespresso/services';
+import type { GeneralSettings, User } from '@eventespresso/data';
+import type { RelationalData } from '@eventespresso/services';
 import type { FormSectionRaw, FormElementRaw } from '@eventespresso/form-builder';
 import type { DatetimeEdge, TicketEdge, PriceEdge, PriceTypeEdge, Event, EventManager, VenueEdge } from './apollo';
